feat: 결제 실패, 펀딩 취소 이벤트 리스너 구현
취소 및 실패된 수량만큼 재고 복구 기능 구현

+++ reward/src/main/java/com/nowayback/reward/application/reward/RewardStockService.java
@@ -121,7 +121,7 @@ public class RewardStockService {
         option.decreaseStock(item.quantity());
         reward.syncStatus();
 
-        Integer itemAmount = option.calculateTotalAmount(item.quantity());
+        Long itemAmount = option.calculateTotalAmount(item.quantity());
 
         StockReservation saveReservation = createAndSaveReservation(
                 fundingId, reward.getId(), option.getId(), item.quantity()
@@ -147,7 +147,7 @@ public class RewardStockService {
     ) {
         reward.decreaseStock(item.quantity());
 
-        Integer itemAmount = reward.calculateTotalAmount(item.quantity());
+        Long itemAmount = reward.calculateTotalAmount(item.quantity());
         StockReservation saveReservation = createAndSaveReservation(
                 fundingId, reward.getId(), null, item.quantity()
         );
